+++ doc/conf.py
@@ -293,7 +293,6 @@ intersphinx_mapping = {
     "dask": ("https://docs.dask.org/en/stable/", None),
     "numba": ("https://numba.pydata.org/numba-doc/latest/", None),
     "numpy": ("https://numpy.org/doc/stable/", None),
-    "pandas": ("https://pandas.pydata.org/pandas-docs/stable/", None),
     "scipy": ("https://docs.scipy.org/doc/scipy/reference/", None),
 }